bigquery_vector_serch/bq_vector_search_api/cloud_run/requirements.txt
uvicorn==0.29.0
fastapi[all]==0.111.0
cloudevents==1.10.1
pypdf==3.16.1
google-cloud-aiplatform>=1.50.0
langchain==0.1.17
pgvector==0.1.8
asyncio==3.4.3
asyncpg==0.27.0
numpy==1.23.2
transformers==4.34.0
langchain-google-vertexai==1.0.3
google-cloud-bigquery==3.21.0
